requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
ecdsa==0.18.0
fastapi==0.110.0
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.14.0
google-cloud-translate==3.15.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
idna==3.6
install==1.3.5
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-google-vertexai==0.0.6
langsmith==0.1.9
marshmallow==3.21.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
passlib==1.7.4
pgvector==0.2.5
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
types-protobuf==4.24.0.20240129
types-requests==2.31.0.20240218
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
yarl==1.9.4
